requirements.txt
# Core ML/NLP - Using flexible versions for Python 3.14 compatibility
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0

# MLOps Tools
mlflow>=2.7.0
dvc>=3.20.0
apache-airflow>=2.7.0

# Monitoring
prometheus-client>=0.17.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
scipy>=1.11.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0